Refactor: Directory 이름 변경
