+++ server/src/configs/index.ts
