+++ server/src/configs/secretKey.ts
